+++ tests/expected/knex-fk.js
@@ -1,0 +1,32 @@
+/**
+ * @param {import('knex').Knex} knex
+ * @returns {Promise<void>}
+ */
+exports.up = function(knex) {
+  // Test
+  return knex.schema.createTable('FK', function(table) {
+  });
+
+  return knex.schema.createTable('users', function(table) {
+    table.increments('id').primary();
+    table.string('name', 32).notNullable();
+  });
+
+  return knex.schema.createTable('orders', function(table) {
+    table.increments('id').primary();
+    table.decimal('amount').notNullable();
+    table.foreign('user_id').references('users.id');
+  });
+};
+
+/**
+ * @param {import('knex').Knex} knex
+ * @returns {Promise<void>}
+ */
+exports.down = function(knex) {
+  return knex.schema.dropTableIfExists('FK');
+
+  return knex.schema.dropTableIfExists('users');
+
+  return knex.schema.dropTableIfExists('orders');
+};